% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_cloupe_path}
\alias{set_cloupe_path}
\title{Set the path to the local cellgeni/cloupe Python package}
\usage{
set_cloupe_path(path)
}
\arguments{
\item{path}{Character. Path to the root of the cloupe Python package.}
}
\value{
The path, invisibly.
}
\description{
The default is \verb{~/Tools/cloupe}. Use this function to point Loupe2R at a
different installation. Changes persist for the R session only.
}
\details{
To make the setting permanent, add to your \code{.Rprofile}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(loupe2r.cloupe_pkg = "/path/to/cloupe")
}\if{html}{\out{</div>}}
}
